Move `GraphElementVertexId` to `subgraph` module
