+++ apps/hash-graph/lib/graph/src/api/rest/utoipa_typedef/subgraph.rs
@@ -11,9 +11,8 @@ pub use self::{
         Vertices,
     },
 };
-use crate::{
-    identifier::GraphElementVertexId,
-    subgraph::{edges::GraphResolveDepths, SubgraphTemporalAxes},
+use crate::subgraph::{
+    edges::GraphResolveDepths, identifier::GraphElementVertexId, SubgraphTemporalAxes,
 };
 
 #[derive(Serialize, ToSchema)]
