+++ apps/hash-graph/lib/graph/src/subgraph/identifier/edge.rs
@@ -1,9 +1,14 @@
 use serde::{Deserialize, Serialize};
+use type_system::url::BaseUrl;
 use utoipa::ToSchema;
 
-use crate::identifier::{
-    knowledge::EntityId,
-    time::{LeftClosedTemporalInterval, VariableAxis},
+use crate::{
+    identifier::{
+        knowledge::EntityId,
+        ontology::OntologyTypeVersion,
+        time::{LeftClosedTemporalInterval, VariableAxis},
+    },
+    subgraph::identifier::OntologyTypeVertexId,
 };
 
 pub trait EdgeEndpoint {
@@ -12,6 +17,19 @@ pub trait EdgeEndpoint {
 
     fn base_id(&self) -> &Self::BaseId;
     fn revision_id(&self) -> Self::RightEndpoint;
+}
+
+impl EdgeEndpoint for OntologyTypeVertexId {
+    type BaseId = BaseUrl;
+    type RightEndpoint = OntologyTypeVersion;
+
+    fn base_id(&self) -> &Self::BaseId {
+        &self.base_id
+    }
+
+    fn revision_id(&self) -> Self::RightEndpoint {
+        self.revision_id
+    }
 }
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, ToSchema)]
